Наседкин Михаил ДЗ №1 v.03 Переделана структура проекта: продюсер вынесен в библиотеку
