+++ homeworks/task1/src/main/resources/producer.properties
@@ -1,0 +1,8 @@
+bootstrap.servers=localhost:9092
+key.serializer=org.apache.kafka.common.serialization.StringSerializer
+value.serializer=ru.sber.serializer.TransactionSerializer
+partitioner.class=ru.sber.util.ProducerPartitioner
+topic=testTopic
+max.block.ms=1000
+acks=1
+max.in.flight.request.per.connection=1
