+++ libs/general-data/src/main/java/ru/sber/config/KafkaConfig.java
@@ -1,0 +1,18 @@
+package ru.sber.config;
+
+import ru.sber.util.PropertiesLoader;
+
+import java.util.Properties;
+
+public class KafkaConfig {
+
+    /**
+     * метод инициирует загрузку конфигурации в файл
+     * @return - экземпляр класса Properties
+     */
+    public static Properties getKafkaProperties(String propertiesFileName) {
+
+        return PropertiesLoader.loadProperties(propertiesFileName);
+    }
+
+}
