+++ libs/general-data/src/main/resources/json/transaction.json
