+++ pom.xml
@@ -2,13 +2,19 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.example</groupId>
+    <groupId>ru.sber</groupId>
     <artifactId>kafka</artifactId>
     <version>1.0-SNAPSHOT</version>
-    <packaging>jar</packaging>
+    <packaging>pom</packaging>
 
     <name>kafka</name>
     <url>http://maven.apache.org</url>
+    <modules>
+        <module>homeworks/task1</module>
+        <module>libs/producer-utils</module>
+        <module>libs/consumer-utils</module>
+        <module>libs/general-data</module>
+    </modules>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -21,6 +27,7 @@
         <common.configuration.version>1.6</common.configuration.version>
         <slf4j.api.version>2.0.16</slf4j.api.version>
         <jackson-datatype-jsr310.version>2.18.2</jackson-datatype-jsr310.version>
+        <jupiter.version>RELEASE</jupiter.version>
     </properties>
 
     <dependencies>
@@ -60,9 +67,9 @@
             <version>${kafka-clients.version}</version>
         </dependency>
         <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>RELEASE</version>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter</artifactId>
+            <version>${jupiter.version}</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
@@ -71,6 +78,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
                 <configuration>
                     <source>8</source>
                     <target>8</target>
